[FIX] base@0.0.0: disable instead of delete invalid views coming from custom module
When a view coming from a custom module is invalid,
we attempted to delete it,
but then if there was inherited views on it,
it raised a foreign key constraint,
because the ID of the view that we attempted to delete
was still referenced in the inherit_id of the inherited views.

There was multiple possible decisions that could be taken:
 - Delete cascade the inherited views,
   but then if they were custom views done in database and not in custom module xml files,
   the data was lost forever,
 - Delete cascade the inherited views if they come from custom module as well,
   and if not just disable the current view,
   a bit complicated to understand when an invalid view
   coming from a custom module was deleted or disabled
 - Just disable the parent invalid view, whether they come from a custom module or not.

I choose the last option, because:
 - Its the easiest to understand,
 - Its the one for which the code is the shorter,
 - The user does not have to do a -u of its custom module to recover its view,
   it just has to read the upgrade report in the discuss app, click
   on the link to jump to the view, and then re-enable it.

The decision took to delete invalid views coming from custom module was
made on the assumption users were doing a -u on their custom module after
receiving their migrated database, and in that case the view would be added back
by this upgrade. In the case the view was simply disabled, the view was not recovered,
because the upgrade would not set it back to active True.

First, this is not necessarily the case,
users do not especially do the -u when they receive their migrated dump,
Second, it was done when there was not yet a migration report to tell
the view was disabled, with a link to directly jump to it.
This is now the case, and it makes it easier to re-enable the custom view
Third, if changes were done manually in these views coming from custom module,
the data was lost forever as well.

That makes a bunch more advantages than cons
to simply disable the view rather than delete it.

opw-2157158
odoo/upgrade#683