[FIX] mail@0.0.0: post the upgrade report at the mail.message register hook
Not on the `ir.ui.view` `_register_hook`.
It was what was intented when this was introduced.
The fact the `_inherit` is on `ir.ui.view` is a typo.

The point is to be able to add in the upgrade report
messages added through the `ir.ui.view` `_register_hook` method.

Before this revision, messages posted through the `ir.ui.view``_register_hook`
were added after the upgrade report was posted,
and therefore these messages were not included in the upgrade report
message.

There was none yet, but I intend to validate the views through
the `_register_hook` of `ir.ui.view`, and therefore messages
will be added to the upgrade report during this `_register_hook`.

Part of odoo/upgrade#1069

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>