[FIX] util: `uniq_tags` for m2o
In standard, tags are only referenced using many2many fields.
But with a customization, e.g. using studio,
a many2one field could reference tags

e.g.
- create a 10.0 database with `project`
- on `project.task`,
- add a custom many2one field `x_type` with as relation `project.tags`,
- create two tasks,
  - one for which you create the tag `Test1`,
  - the second for which you create the tag `test1`
Then upgrade to 11.0.

Before this revision, the upgrade fails with:
```
  File "/src/odoo/11.0/odoo/addons/base/maintenance/migrations/project/10.saas~15.1.1/pre-uniqtags.py", line 5, in migrate
    util.uniq_tags(cr, 'project.tags', 'lower(name)')
  File "/src/odoo/11.0/odoo/addons/base/maintenance/migrations/util.py", line 1209, in uniq_tags
    assert da == "c"  # should be a ondelete=cascade fk
AssertionError
```

This is not the first case where the issue happens,
but before it was solved with specific scripts:
 - odoo/migration-scripts@e7faed991378dedd056ca4d43ef82d74b8c43521
 - odoo/migration-scripts@f22360a44f68b4d54c7710becfb3ffe8984a192c
 - odoo/migration-scripts@80e242520410b16910ee34ca259fcbcb95e73ee3
 - odoo/migration-scripts@ab07928f8ba830aa9d224e999a5f6db7fb2a2fa2

In addition, during my analysis of this issue,
I came accross another issue searching the specific scripts for
related issues:

The `ON DELETE CASCADE` is not always correctly set
in the relation table of the many2many.
And sometimes its on standard table, sometimes on custom tables
 - odoo/migration-scripts@f5352d1d9ac1d9f4c2e188eb565bb1b7ff4081b3
 - odoo/migration-scripts@b12b07c13dd9636a9eee83235b6aa3cbfc784ea0
 - odoo/migration-scripts@bef6ade9e6b9ee5ae16dd8a2c2d60610918550ed
 - odoo/migration-scripts@cdfa54dbbd458e5fbadbb85159be9fe7baf223a6
 - odoo/migration-scripts@b85122f9b5c6b698a326f6c3665106825419da24
 - odoo/migration-scripts@6d9b89b01cd7f786674b2b2ce7832dc60048abaa
 - odoo/migration-scripts@fc63579d32434acb0beeb9a532fe7a56a33aee09
 - odoo/migration-scripts@41bba45fc966b7c8faf155d114218f385b44a74c
 - odoo/migration-scripts@5133abd696976bd2f96265c7e5f65ba32034f538
 - odoo/migration-scripts@6b4ca6d23fca4f2b414a0de02e7ac3b2c1c472c7
 - odoo/migration-scripts@442b6ded2b4e4b80c9393778a87b29c1a83f474b
 - odoo/migration-scripts@38871939f805283bcd95fdb16346bc49e057ea00
 - odoo/migration-scripts@63c030d935b0fc908d58b4558a92b483a38794fe

When this happens, we add a fallback, checking `ir_model_fields`
to determine if it's a many2many. The fact the `ON DELETE CASCADE`
is missing doesn't really matter for the script to correctly works,
and it's not really its role to ensure the many2many table columns
are correctly set with `ON DELETE CASCADE`.

We could add a script ensuring the default many2many tags table
are correctly set with `ON DELETE CASCADE`,
but this would not cover the custom many2many tables.

Request 53772
M1410231178259

Part of odoo/upgrade#1763

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>