[FIX] util: rename existing fields with the same name as standard fields
It has been a recurring error that during an upgrade a field is renamed
and the upgrade crashes because a field with the same name already
existed in the model, coming from a custom module.

The fix usually involves writing a custom upgrade script and renaming
the existing fields by appending `_old` or `_custom`.

This commit attempts to fix this in a general way, by appending the
`_custom` to the existing field and all of its references.

Part of odoo/upgrade#1812

Signed-off-by: Christophe Simonis <chs@odoo.com>