[FIX] util: specify utf-8 encoding on migration reports
In Python 3, the default encoding for the `decode` method is `utf-8`[1].
In Python 2 however, the default encoding is `ascii`[2].

In the upgrade scripts, the migration reports message is obtained using
the `decode` method with default arguments. This can cause a problem if
the database is upgrading from an older Odoo version which used Python 2
(for example 9.0 to 10.0) and the message contains `utf-8` characters,
as it will try to decode it as `ascii`.

To avoid this problem, we specify the `utf-8` encoding explicitly in the
`decode` method.

Request ID: 54409
opw: 2345895
[1] https://docs.python.org/3.6/library/codecs.html#codecs.decode
[2] https://docs.python.org/2.7/library/codecs.html#codecs.decode

Part of odoo/upgrade#1827

Signed-off-by: Christophe Simonis <chs@odoo.com>