[REM] util.disable_triggers
The point of this method was to deactivate the triggers (FK) that are
executed by PostgreSQL on the second update on a table.
It was used on tables that can contains a lot of records, and thus have
a slow verification of constraints.

First, this method only works when the connected PostgreSQL user is
superuser. This not the case (for security reasons) on the new upgrade
platform.

Second, this postgresql behavior can be bypassed by executing the
queries in parallel as it start by commiting, which restart a new
transation. It's also faster on big tables.

Part of odoo/upgrade#1963

Signed-off-by: Christophe Simonis <chs@odoo.com>