[FIX] util._update_view_key: consider cow views
This part of the code was not considering COW views that should also receive
the `key` update (which is sync'd with its new `xmlid`).

Also, a view `key` is not always the same as its `xmlid`, this is typically the
case for views created through a `website.page` record in XML. The view will
then have an `xmlid` suffixed by `_view`.

task-2462993
Part of odoo/upgrade#2452