[FIX] util/jinja_to_qweb.py: make user available in the jinja rendering context
Add the user variable for jinja rendering that need to use it.
Otherwise, it is a source of difference between the render in jinja to
qweb. Wich then would lead to a false negative when asserting if a
template has been properly migrated.

Oversight of bfb1ca67b2207b5139ef68398193c118a17bd1aa

Part-of: odoo/upgrade#3019