[IMP] util/specific: add helper to rename table/columns
This is a very common task that most specific scripts use in a one-shot
fashion. It's repetitive enough that it's annonying. It is also error
prone: forget to check the table/column exists and the upgrade fails due
to the specific script. The message to customers is in many cases also
mispelled and confusing.

Part of odoo/upgrade#3154

Signed-off-by: Christophe Simonis <chs@odoo.com>