[FIX] base/0.0.0: Exclude `moved0` field from `_backup` tables
When upgrading a database which have some `moved0` fieldsat start, the
upgrade will lead to a crash.
To explain this, let's take an account script as an example:
- `base/0.0.0/pre-moved0.py` is run first and will take care to add in
`ENVIRON["moved0"]` every tuple of `('table', 'field')` where the field
is a `moved0` one.
- `account/saas~13.4.1.1/pre-10-payment-refactoring.py` is run in
second and will create some `_backup` table.
- `base/0.0.0/end-moved0.py` is finally run in third and will recover
every tuple of `('table', 'field')` where the fields is a `moved0` one
(at the end of upgrade). If there is some new `moved0` field, these
will be removed only if the column is empty. If not empty, an error is
raised.

upg-106121
upg-106591
upg-104274 (solved with a specific script)

```
Traceback (most recent call last):
  File "/home/odoo/src/odoo/14.0/odoo/service/server.py", line 1199, in preload_registries
    registry = Registry.new(dbname, update_module=update_module)
  File "/home/odoo/src/odoo/14.0/odoo/modules/registry.py", line 89, in new
    odoo.modules.load_modules(registry._db, force_demo, status, update_module)
  File "/home/odoo/src/odoo/14.0/odoo/modules/loading.py", line 475, in load_modules
    migrations.migrate_module(package, 'end')
  File "/home/odoo/src/odoo/14.0/odoo/modules/migration.py", line 180, in migrate_module
    migrate(self.cr, installed_version)
  File "/tmp/tmpuo_ygaw_/migrations/base/0.0.0/end-moved0.py", line 30, in migrate
    % "\n".join("\t- %s.%s" % m for m in sorted(remaining_fields))
odoo.upgrade.util.exceptions.MigrationError: New `moved0` field. It happen when the ORM cannot change a column type by itself.
	- account_payment_pre_backup.writeoff_mode_moved0
```

Part of odoo/upgrade#3243

Signed-off-by: Christophe Simonis <chs@odoo.com>