[FIX] util/modules: _force_upgrade_of_fresh_module is one level down
Since this function is now called a level deeper on the stack we get as
version `'util'` which is incorrect.

The solution is to get the information from the upper level in the
frame.

The breaking change comes from odoo/upgrade#3308

upg-166178
(others are also affected)

Part of odoo/upgrade#3363

Signed-off-by: Christophe Simonis <chs@odoo.com>
Co-authored-by: Christophe Simonis <chs@odoo.com>