[FIX] pre-commit: bump black version to 22.3.0
Before this commit
The black pre-commit step is broken due to an incompatibility
with a new version of one of its dependencies (click 8.1+).
See: https://github.com/psf/black/issues/2964#issuecomment-1080974737

After this commit
The new version does not have this incompatibility.
This commit bumps the black version to 22.3.0.

Part of odoo/upgrade#3391

Signed-off-by: Christophe Simonis <chs@odoo.com>