[FIX] util/accounting: fix upgrade_tax_tags
The function 'create' is overriden in the upgrade repo. When a record already
exists with the same fields (for a tax: name, type_tax_use, company_id,
tax_scope), a new one is not created ! Instead, the existing record is
retrieved and some other fields are written on the existing record directly.

It causes a traceback when we have an existing tax, similar to a template
tax, but the existing tax is linked to account_move_lines. Indeed, rather
than creating a new tax, it is retrieved and we try to create the
repartition_lines on this existing tax (the query begins with (5, 0, 0)).
It then crashes because we can't unlink repartition_lines when they are
linked to amls.

The fix is simply to rename the existing taxes that are similar to template
taxes before calling 'create_record_with_xmlid'.

Part of odoo/upgrade#3562

Signed-off-by: Christophe Simonis <chs@odoo.com>