[FIX] flake8: ignore E203 to avoid whitespace issue
Matt happens to go red due to the changes made by black
which will enforce an equal amount of whitespace around
slice operators. Due to this Flake8 will raise E203
whitespace before ':' warnings. Since, this warning is
not PEP8 compliant, Fix ignores it to be checked

Part of odoo/upgrade#3964

Signed-off-by: Christophe Simonis <chs@odoo.com>