[IMP] util/snippets.py: abstract pre-pickling
Take the necessary steps to make the `html_converter` callbacks
pickleable in `snippets.py`, abstracting it from migration scripts.

Now that `HTMLConverter` is used, `html_selector_converter` can be
merged into `html_converter`, getting the same result if the latter is
called with the `selector` argument.

Followup on 959ff243

Part of odoo/upgrade#4241

Signed-off-by: Christophe Simonis <chs@odoo.com>