[IMP] util/inconsistencies.py: Add `verify_{uoms,products}` functions
UoM issues in upgrade are not always quite clear: the same error can
happen in multiple places and then is not everytime easy to understand.
The issue is the following:

```
The unit of measure %s defined on the order line doesn't belong to the
same category as the unit of measure %s defined on the product. Please
correct the unit of measure defined on the order line or on the
product, they should belong to the same category.
```

To improve the standard uom issue message, two new functions has been
created:
- `verify_uoms`: This method will ensure that every uom
defined on table lines match the same category than which is defined
on the product template. If some inconsistencies are found, an error
will be raised to list every line that have mismatching uom category.
- `verify_products`: This method will ensure that every
product defined on table lines match the product defined on the
foreign_table. If some inconsistencies are found, an error will be
raised to list every line that have mismatching product.

upg-470381
upg-462352
upg-437441
upg-418532
(Lots of request have failed on it, about 500+ taking achived ones into
account.)

tbg-71

Part-of: odoo/upgrade#4066