[ADD] util/pg: add a specialized `create_fk` helper
This is useful when creating columns that are meant to be used as FK. By
setting the FK constrains right at creation time we may influence PG
planner positively for the rest usage of this column in subsequence
queries. Otherwise any query run before the ORM creates the FK
constraint may be badly planned.

Part-of: odoo/upgrade#4556