[IMP] util/record.py: regard noupdate in ensure_xmlid_match_record
The change in f8c41cc7a4c8314dc31c6415a98be512c90c522c introduced additional
`ir_model_data` updates for rows that point to an existing record that doesn't
match the values used to lookup a suitable record, if such a matching record
exists. It will update the row in that case not regarding the `noupdate`
column.

Tune this behavior by adding an additional check that returns immediately with
the original `res_id`, if an `ir_model_data` row for the xmlid exists that
points to an existing record and the `noupdate` column is `True`. This way, the
update for existing, non-matching records will only happen if `noupdate` is
`False`.