[IMP] util/pg.py: Handle inheritance for id sequence creation
For the moment, when a sequence is missing, it will be created using
the sequence name `{table_name}_id_seq`.
But in case of inheritance, we actually use the table name of a child
to create sequence instead of using the parent (or master) table name.

For example, these following tables herits from `ir_actions` table and
must use the sequence associated to `ir_actions`.
```
+-------------------+-------------------+
| table_name        | sequence_name     |
+-------------------|-------------------|
| ir_actions        | ir_actions_id_seq |
| ir_act_client     | ir_actions_id_seq |
| ir_act_report_xml | ir_actions_id_seq |
| ir_act_server     | ir_actions_id_seq |
| ir_act_url        | ir_actions_id_seq |
| ir_act_window     | ir_actions_id_seq |
+-------------------+-------------------+
```

So we have to make sure that when we set up a sequence, the sequence is
well created for the parent and not for the children.

Note: This only handle simple inheritance (One child have only one
parent, and his parent (if applicable), also striclty have one parent)
So it's safe to take the last parent as we go up from child to parent.
The last will ever be the master parent.

The multiple inheritance (A child herit from multiple parents) is not
implemented, this is because it doesn't exists in Odoo's db for now.

Part of odoo/upgrade#4952

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>