[IMP] util.alter_column_type: cleanup selection values
If the column was a selection field, we should explicitly remove the
`ir_model_fields_selection` records as their `unlink` will try to reset
the selection value to another varchar value.
However, the column is no more a varchar (else we didn't called
`alter_column_type` to start with).

Part of odoo/upgrade#5080

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>