[FIX] util/records.py: Fix `remove_record` premature xmlid deletion
The `remove_record` function was exhibiting unexpected behavior by
prematurely deleting the xmlid before passing control to `remove_view`,
resulting in the failure to remove COW (Copy-On-Write) views because the
xmlid was absent.

To address this issue, the commit updates the query to retrieve the data
instead of directly deleting it. This adjustment then leaves the
responsibility of xmlid removal to the delegated methods or the
`remove_records` function.

Part of odoo/upgrade#5089

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>