[FIX] util.adapt_domains: don't update unmodified views
The `util.edit_view` contextmanager always write back the view, even if
it hasn't been modified.
This cause an issue with the saas-16.1 script that remove the
`__last_update` field on all model as it is done using multi processes.
If any view contains the `__last_update` field, as multiple processes
try to modify this view, a `psycopg2.errors.SerializationFailure`
exception is raised.

tgb-881

Part of odoo/upgrade#5160

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>