[FIX] util.remove_field: keep field info
When removing a field, we can now keep the tracking as 'field_id' is not
required anymore and tracking model holds a 'fields_info' column to keep
removed field information.

This globally matches the behavior implemented at `IrModelField.unlink()`
as done in odoo/odoo#124182 . Note that tracking sequence is lost as it is
not stored in DB (it is a field parameter available in registry) but it
seems like an acceptable trade-off.

Task-3345979

Part of odoo/upgrade#5239

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>