[FIX] base/tests: correctly test `util.update_record_from_xml`
Test the behavior of the `ensure_references` argument in its own
function.

Trying to update `base.menu_security` with the `ensure_references` flag
set will lead to force the update of the `base.main_company` record.
This however fail when a localization is installed as it will try to
update the company's currency, which is forbidden as some journal items
already exist.

Part of odoo/upgrade#5256

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>