[FIX] util.trigger_auto_install
Correctly set the auto_install modules as installed.
Just changing the state is not enough. All its dependencies should be
installed.

As example, in version 16.0, the new `spreadsheet_dashboard_account` module
is defined as
```
{
    'depends': ['spreadsheet_dashboard', 'account'],
    'auto_install': ['account'],
}
```

So, when `account` is installed, this new module should be auto
installed, but also its uninstalled dependencies. In this case, the
module `spreadsheet_dashboard` should also be marked as installed, as
well as its dependencies, recursively.

Part of odoo/upgrade#5343

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>