[FIX] util/orm.py: Restore compatibility with 7.0
When the module `runsanitizer` is loaded, Odoo is trying to import `env()`
from util/orm.py. There is a check to ensure that `openerp.fields`, which
was introduced in 8.0, is not being read by earlier versions and set to None
instead.

While parsing orm.py however, all decorators are evaluated including
`no_selection_cache_validation()`, which assumes that `openerp.fields` has
been imported, and fails on an AttributeError. A check at the top of the
decorator fixes that (it was already anticipated that the decorated function
`recompute_fields` would not work in 7.0).