[IMP] util/{records,domains}: improve edit_view context manager
The generator based context manager, using `yield`, has multiple
shortcomings. First it requires an extra `skippable_cm` magic to be able
to deal with missing views. Second, there was not clear way to signal
that we want to skip the edition of the view to avoid spurious writes if
in the end the view is unchanged.