[FIX] records: skip inactive, invalid views in adapt_domains()
Do not fail when the XML parser fails to parse an inactive view. Inactive views
are probably abandoned and do not cause immediate errors in the upgraded db, so
just skip them when they are invalid.

Also, log the view_id before (re)raising a parser exception. This should save
some time when manual fixing of invalid (active) views is needed.

upg-1171125