[FIX] util/pg: ensure the ranges contain at least bucket_size elements
Before this patch when we exploded a query by range we could end up
creating too many queries for which there is no actual record in the
range. In extreme cases this causes a memory error doe to the huge list
of generated queries.

Example:
```
=> select min(id),max(id),count(id) from account_journal
+-----+----------+-------+
| min | max      | count |
|-----+----------+-------|
| 1   | 10000003 | 56    |
+-----+----------+-------+
```

All the journals could be processed in one bucket of >=56 elements. The
high max id causes the generation of many queries that would not process
any actual record.

```
...
  File "/tmp/tmpl7nu4jh5/migrations/account/saas~16.2.1.2/pre-migrate.py", line 365, in migrate
    util.explode_execute(cr, query, table="account_journal", alias="journal", bucket_size=1)
  File "/tmp/tmpl7nu4jh5/migrations/util/pg.py", line 226, in explode_execute
    explode_query_range(cr, query, table, alias=alias, bucket_size=bucket_size),
  File "/tmp/tmpl7nu4jh5/migrations/util/pg.py", line 217, in explode_query_range
    return [
  File "/tmp/tmpl7nu4jh5/migrations/util/pg.py", line 218, in <listcomp>
    cr.mogrify(query, {"lower-bound": index, "upper-bound": index + bucket_size - 1}).decode()
  File "/home/odoo/src/odoo/17.0/odoo/sql_db.py", line 316, in mogrify
    return self._obj.mogrify(query, params)
MemoryError
```

On this patch we propose to fallback to precise bucketing. Ensuring each
bucket has the same number of records from the table -- with the
possible exception of the last bucket.

closes odoo/upgrade-util#56

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>