[FIX] util/snippets: generate HTML instead of XML for mailing.mailing
The `convert_html_content` tool writes XML values if a change happened.
This is a problem for the fields of `mailing.mailing` because they
strictly expect HTML content.
E.g. if `<img src="...">` becomes `<img src="..."/>`, the wysiwyg
editor displays an empty template instead of the actual content.

This commit avoids this by making the writes to the fields of
`mailing.mailing` formatted in HTML.

task-3797854