[IMP] base/17.0: tools for converting view attrs
Since Odoo 17.0 we no longer allow for domains in `attrs` attributes of
view elements. Domains were converted to Python expressions. In this
commit we add some tooling for converting them.

The main entry point is the function `fix_attrs` -- see docstring for
details. An example usage function is `adapt_view`

Closes #54