[FIX] util/records: update_record_from_xml - reset all translations
This commit modifies `update_record_from_xml()` to ensure that translations are updated in _all_ languages by adding the `overwrite` parameter when calling `_update_translations()`.

Without this parameter, existing translations are only overwritten if they are identical to the `en_US` translation, which is not a true reset. This can be a particular issue with mail templates.

It also handles Odoo versions correctly.