[IMP] specific: check if custom model exists
When a custom model conflicts with a standard one we may choose to write a
specific fix, using the `rename_custom_model` util to rename it.

However, although unlikely, customers may remove that model in successive
upgrade requests, making the call to `rename_model` unnecessary and the extra
entry in the migration report wrong.