[IMP] specific: check if custom module exists
When a custom module conflicts with a standard one we may choose to write a
specific fix, using the `rename_custom_module` util to rename it.

However, although unlikely, customers may remove that module in successive
upgrade requests, making the call to `rename_module` unnecessary and the extra
entry in the migration report wrong.