[IMP] specific: check if custom table exists
When a custom table conflicts with a standard one we may choose to write a
specific fix, using the `rename_custom_table` util to rename it.

However, customers may remove that table in successive upgrade requests, making
the call to `rename_table` unnecessary and the extra entry in the migration
report wrong.

Part-of: odoo/upgrade-util#91